requirements.txt
gluoncv
mxnet
torch
torchvision
matplotlib.pyplot
numpy
scipy
opencv